backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
anthropic>=0.49.0
python-dotenv==1.1.0
pydantic==2.11.4
